+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidvci/CredentialSchemeMapping.kt
@@ -4,6 +4,8 @@ import at.asitplus.openid.ClaimDescription
 import at.asitplus.openid.CredentialFormatEnum
 import at.asitplus.openid.CredentialFormatEnum.DC_SD_JWT
 import at.asitplus.openid.CredentialFormatEnum.JWT_VC
+import at.asitplus.openid.OpenId4VciClaimsPathPointer
+import at.asitplus.openid.OpenId4VciClaimsPathPointerSegmentString
 import at.asitplus.openid.OpenIdConstants.BINDING_METHOD_COSE_KEY
 import at.asitplus.openid.OpenIdConstants.BINDING_METHOD_JWK
 import at.asitplus.openid.OpenIdConstants.URN_TYPE_JWK_THUMBPRINT
@@ -64,7 +66,9 @@ fun CredentialScheme.toIsoMdocSupportedCredentialFormat(identifier: String): Pai
         docType = isoDocType!!,
         supportedBindingMethods = setOf(BINDING_METHOD_JWK, BINDING_METHOD_COSE_KEY),
         isoClaims = claimNames.map {
-            ClaimDescription(path = listOf(isoNamespace!!) + it.split("."))
+            ClaimDescription(path = OpenId4VciClaimsPathPointer(isoNamespace!!) + it.split(".").map {
+                OpenId4VciClaimsPathPointerSegmentString(it)
+            })
         }.toSet()
     )
 
@@ -76,7 +80,7 @@ fun CredentialScheme.toPlainJwtSupportedCredentialFormat(identifier: String): Pa
         ),
         supportedBindingMethods = setOf(BINDING_METHOD_JWK, URN_TYPE_JWK_THUMBPRINT),
         vcJwtClaims = claimNames.map {
-            ClaimDescription(path = it.split("."))
+            ClaimDescription(path = OpenId4VciClaimsPathPointer(it.split(".")))
         }.toSet()
     )
 
@@ -86,7 +90,7 @@ fun CredentialScheme.toSdJwtSupportedCredentialFormat(identifier: String): Pair<
         sdJwtVcType = sdJwtType!!,
         supportedBindingMethods = setOf(BINDING_METHOD_JWK, URN_TYPE_JWK_THUMBPRINT),
         sdJwtClaims = claimNames.map {
-            ClaimDescription(path = it.split("."))
+            ClaimDescription(path = OpenId4VciClaimsPathPointer(it.split(".")))
         }.toSet()
     )