+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/SameScopeCredentialSchemeMapper.kt
@@ -4,6 +4,7 @@ import at.asitplus.openid.ClaimDescription
 import at.asitplus.openid.CredentialFormatEnum
 import at.asitplus.openid.CredentialFormatEnum.DC_SD_JWT
 import at.asitplus.openid.CredentialFormatEnum.JWT_VC
+import at.asitplus.openid.OpenId4VciClaimsPathPointer
 import at.asitplus.openid.OpenIdConstants.BINDING_METHOD_COSE_KEY
 import at.asitplus.openid.OpenIdConstants.BINDING_METHOD_JWK
 import at.asitplus.openid.OpenIdConstants.URN_TYPE_JWK_THUMBPRINT
@@ -48,7 +49,7 @@ class SameScopeCredentialSchemeMapper(
                     docType = isoDocType!!,
                     supportedBindingMethods = setOf(BINDING_METHOD_JWK, BINDING_METHOD_COSE_KEY),
                     isoClaims = claimNames.map {
-                        ClaimDescription(path = listOf(isoNamespace!!) + it.split("."))
+                        ClaimDescription(path = OpenId4VciClaimsPathPointer(listOf(isoNamespace!!) + it.split(".")))
                     }.toSet()
                 )
             }
@@ -62,7 +63,7 @@ class SameScopeCredentialSchemeMapper(
                     ),
                     supportedBindingMethods = setOf(BINDING_METHOD_JWK, URN_TYPE_JWK_THUMBPRINT),
                     vcJwtClaims = claimNames.map {
-                        ClaimDescription(path = it.split("."))
+                        ClaimDescription(path = OpenId4VciClaimsPathPointer(it.split(".")))
                     }.toSet()
                 )
             }
@@ -74,7 +75,7 @@ class SameScopeCredentialSchemeMapper(
                     sdJwtVcType = sdJwtType!!,
                     supportedBindingMethods = setOf(BINDING_METHOD_JWK, URN_TYPE_JWK_THUMBPRINT),
                     sdJwtClaims = claimNames.map {
-                        ClaimDescription(path = it.split("."))
+                        ClaimDescription(path = OpenId4VciClaimsPathPointer(it.split(".")))
                     }.toSet()
                 )
             }